move impl to relay.rs
