+++ src/rpc/relay.rs
@@ -10,7 +10,8 @@ use crate::{
     provider::ProviderExt,
     rpc::ExtraFeeInfo,
     signers::Eip712PayLoadSigner,
-    transactions::interop::InteropBundle,
+    storage::BundleHistoryEntry,
+    transactions::{RelayTransactionKind, interop::InteropBundle},
     types::{
         Account, Asset, AssetDeficit, AssetDiffResponse, AssetMetadataWithPrice, AssetPrice,
         AssetType, Call, ChainAssetDiffs, DelegationStatus, Escrow, FundSource,
@@ -20,11 +21,12 @@ use crate::{
         Quotes, SignedCall, SignedCalls, SourcedAsset, Transfer, VersionedContracts,
         rpc::{
             AddFaucetFundsParameters, AddFaucetFundsResponse, AddressOrNative, Asset7811,
-            AssetFilterItem, CallKey, CallReceipt, CallStatusCode, ChainCapabilities,
-            ChainFeeToken, ChainFees, GetAssetsParameters, GetAssetsResponse,
-            GetAuthorizationParameters, GetAuthorizationResponse, GetCallsHistoryParameters,
-            GetCallsHistoryResponse, Meta, PreCallContext, PrepareCallsCapabilities,
-            PrepareCallsContext, PrepareUpgradeAccountResponse, RelayCapabilities, RequiredAsset,
+            AssetFilterItem, CallHistoryCapabilities, CallHistoryEntry, CallHistoryTransaction,
+            CallKey, CallReceipt, CallStatusCode, ChainCapabilities, ChainFeeToken, ChainFees,
+            GetAssetsParameters, GetAssetsResponse, GetAuthorizationParameters,
+            GetAuthorizationResponse, GetCallsHistoryParameters, GetCallsHistoryResponse, Meta,
+            PreCallContext, PrepareCallsCapabilities, PrepareCallsContext,
+            PrepareUpgradeAccountResponse, RelayCapabilities, RequiredAsset,
             SendPreparedCallsCapabilities, SortDirection, UpgradeAccountContext,
             UpgradeAccountDigests, ValidSignatureProof,
         },
@@ -3091,14 +3093,149 @@ impl RelayApiServer for Relay {
             }
         };
 
-        // Fetch entries from storage (all logic moved to storage layer)
-        let entries = self
+        // Fetch bundles from underlying storage
+        let bundles = self
             .inner
             .storage
-            .get_calls_history(params.address, params.limit, offset, sort_desc)
+            .get_bundles_by_address(params.address, params.limit, offset, sort_desc)
             .await?;
 
-        Ok(entries)
+        let entries_futures = bundles.into_iter().enumerate().map(async |(idx, history_entry)| {
+            let index: u64 = offset + idx as u64;
+
+            let entry = match history_entry {
+                // Multi-chain bundle path
+                BundleHistoryEntry::Interop { bundle: bundle_with_status, timestamp } => {
+                    let bundle = &bundle_with_status.bundle;
+                    let status = bundle_with_status.status;
+
+                    // Extract key hash from first destination transaction signature
+                    // For 65-byte EOA signatures, use 0x0 as keyHash
+                    let key_hash = bundle
+                        .dst_txs
+                        .first()
+                        .and_then(|tx| match &tx.kind {
+                            RelayTransactionKind::Intent { quote, .. } => {
+                                if quote.intent.signature().len() == 65 {
+                                    // Signed by the root eoa key.
+                                    Some(B256::ZERO)
+                                } else {
+                                    Signature::decode_key_hash(quote.intent.signature())
+                                }
+                            }
+                            _ => None,
+                        })
+                        .unwrap_or(B256::ZERO);
+
+                    // Get transaction statuses using batch method
+                    let all_tx_ids: Vec<_> = bundle
+                        .src_txs
+                        .iter()
+                        .chain(bundle.dst_txs.iter())
+                        .map(|tx| tx.id)
+                        .collect();
+
+                    let tx_statuses =
+                        self.inner.storage.read_transaction_statuses(&all_tx_ids).await?;
+
+                    let mut transactions = Vec::with_capacity(tx_statuses.len());
+                    for tx_status_opt in tx_statuses {
+                        if let Some((chain_id, tx_status)) = tx_status_opt
+                            && let Some(tx_hash) = tx_status.tx_hash()
+                        {
+                            transactions.push(CallHistoryTransaction {
+                                chain_id,
+                                transaction_hash: tx_hash,
+                            });
+                        }
+                    }
+
+                    // Extract quotes from bundle (both src and dst transactions)
+                    let quotes = bundle
+                        .src_txs
+                        .iter()
+                        .chain(bundle.dst_txs.iter())
+                        .filter_map(|tx| match &tx.kind {
+                            RelayTransactionKind::Intent { quote, .. } => Some((**quote).clone()),
+                            _ => None,
+                        })
+                        .collect::<Vec<_>>();
+
+                    // TODO: Implement asset diff
+                    let asset_diff = AssetDiffResponse::default();
+
+                    CallHistoryEntry {
+                        id: bundle.id,
+                        index,
+                        status: status.to_call_status_code(),
+                        timestamp,
+                        transactions,
+                        key_hash,
+                        capabilities: CallHistoryCapabilities { asset_diff, quotes },
+                    }
+                }
+
+                // Single-chain bundle path
+                BundleHistoryEntry::SingleChain {
+                    bundle_id,
+                    chain_id,
+                    quote,
+                    tx_hash,
+                    timestamp,
+                } => {
+                    let (key_hash, quotes) = if let Some(quote) = quote {
+                        let key_hash = if quote.intent.signature().len() == 65 {
+                            // Signed by root eoa key
+                            B256::ZERO
+                        } else {
+                            Signature::decode_key_hash(quote.intent.signature())
+                                .unwrap_or(B256::ZERO)
+                        };
+                        (key_hash, vec![(*quote)])
+                    } else {
+                        // Old transaction without stored quote data
+                        (B256::ZERO, vec![])
+                    };
+
+                    // Build transaction list (only if tx_hash exists)
+                    let transactions = tx_hash
+                        .map(|tx_hash| {
+                            vec![CallHistoryTransaction { chain_id, transaction_hash: tx_hash }]
+                        })
+                        .unwrap_or_default();
+
+                    // TODO: Implement asset diff
+                    let asset_diff = AssetDiffResponse::default();
+
+                    // Get tx_id from bundle_transactions table
+                    let tx_ids = self.inner.storage.get_bundle_transactions(bundle_id).await?;
+                    let call_status = if let Some(&tx_id) = tx_ids.first() {
+                        self.inner
+                            .storage
+                            .read_transaction_status(tx_id)
+                            .await?
+                            .map(|(_, tx_status)| tx_status.to_call_status_code())
+                            .unwrap_or(CallStatusCode::Pending)
+                    } else {
+                        CallStatusCode::Pending
+                    };
+
+                    CallHistoryEntry {
+                        id: bundle_id,
+                        index,
+                        status: call_status,
+                        timestamp,
+                        transactions,
+                        key_hash,
+                        capabilities: CallHistoryCapabilities { asset_diff, quotes },
+                    }
+                }
+            };
+
+            Ok::<_, StorageError>(entry)
+        });
+
+        Ok(try_join_all(entries_futures).await?)
     }
 
     async fn verify_signature(
